+++ lib/src/davi.dart
@@ -21,9 +21,6 @@ import 'package:flutter/services.dart';
 /// Table view designed for a large number of data.
 ///
 /// The type [DATA] represents the data of each row.
-/// The [cellContentHeight] is mandatory due to performance.
-/// The total height of the cell will be the sum of the [cellContentHeight]
-/// value, divider thickness, and cell margin.
 class Davi<DATA> extends StatefulWidget {
 //TODO handle negative values
 //TODO allow null and use defaults?
@@ -41,6 +38,7 @@ class Davi<DATA> extends StatefulWidget {
       int? visibleRowsCount,
       this.focusable = true,
       this.multiSort = false,
+      this.tapToSortEnabled = true,
       this.lastRowWidget,
       this.rowColor,
       this.rowCursor,
@@ -67,6 +65,9 @@ class Davi<DATA> extends StatefulWidget {
   final bool multiSort;
   final Widget? lastRowWidget;
   final OnLastRowWidgetListener? onLastRowWidget;
+
+  /// Indicates whether sorting events are enabled on the header.
+  final bool tapToSortEnabled;
 
   @override
   State<StatefulWidget> createState() => _DaviState<DATA>();
@@ -181,6 +182,7 @@ class _DaviState<DATA> extends State<Davi<DATA>> {
         child: TableLayoutBuilder(
             onHover: widget.onHover != null ? _setHoveredRowIndex : null,
             multiSort: widget.multiSort,
+            tapToSortEnabled: widget.tapToSortEnabled,
             scrollControllers: _scrollControllers,
             columnWidthBehavior: widget.columnWidthBehavior,
             themeMetrics: themeMetrics,